[Android] Added bbque_service Android Library
The bbque_service library provides the API to develop AEM-integrated
Android applications.

The library can be imported as a app module through Android Studio to
develop applications.
It defines the BarbequeService class, which has to be extended by the
developer in order to implements the application specific service.
The user-definde application's Activity has to bind this Service
in order to exploit RTLib and BarbequeRTRM functionalities.
The base class provide the definition of a base message Handler to
manage incoming message from the application Activity. The developer can
extends the handler with its own message.
The BarbequeService base class contains the definition and the instance
of the AndroidBbqueEXC, which in turn extends the BbqueEXC of the
Java wrapper to exploit RTLib native functions.
The AndroidBbqueEXC contains a basic implementation of the AEM
callbacks.
The developer, extending the base AndroidBbqueEXC class, has to override
in order to implement the applications functionalities.

Signed-off-by: Michele Zanella <michele.zanella@polimi.it>