[RA] Power management member functions
These functions are expected to be called by the scheduling policies
(EnqueueResourceToPowerManage) to specify power management actions to
be applied to specific resources. Such actions can consist of turning
on/off, change frequency or performance state and so on...

Once the resource assignment is enforced, the daemon can perform all the
pending resource management actions, by calling
DequeueResourceToPowerManage(). This will return a ResourcePtr_t for
each call, until all the resources have been popped out. For each
resource object we can get the current power settings and check if
there are pending actions to apply.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>