[PP/PLM] ActuatePowerManagement: refactoring
Added an overloaded function to process a specific resource.
The void function is expected to include actions involving not managed
resources.

The PlatformManager implementation fetches the pending resources from the
ResourceAccounter and dispatch them to the specific proxy, depending
on the resource location (remote or local)

The LocalPlatformProxy performs the suitable actions by calling the
PowerManager.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>