[SchedPol_Test] Completely new task-graph mapping approach
Using the recently introduced functions of ResourceAccounter for mapping
the tasks onto the first available processors, following the preference
order specified with the 'hw_prefs' attribute of the recipe file.

The new WorkingMode function are then used to transparently fill the object
with resource requests and for each of them perform the binding.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>