[EYM/WM/TRIG] Deep revision of the triggers
The Power Monitor is responsible for the "temperature" trigger, while the
Energy Monitor is for the "battery" (charge level and current drain) ones.

The Trigger objects have been revised to directly call the action function
after performing the specific Check() function, in case of return value
equal to true. This call to Check() is performed once a new updated value
is evaluated through the NotifyUpdatedValue() function.

The triggers are loaded only if the related line of trigger type, in the
configuration file, has been uncommented.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>